requirements.txt
openai
foundry-local-sdk
httpx
textual